requirements.txt
discord.py==2.3.2
requests==2.31.0
SQLAlchemy==2.0.23